computeIntent

+++ src/main/java/com/accenture/gameoftanks/core/Intent.java
@@ -3,7 +3,7 @@ import java.lang.Math;
 public class Intent {
     private boolean onMove;
     private boolean onShot;
-    private double moveAngle;
+    private float moveAngle;
     private float shotAngle;
 
     //Generate constructor
@@ -35,7 +35,7 @@ public class Intent {
         return moveAngle;
     }
 
-    public void setMoveAngle(double moveAngle) {
+    public void setMoveAngle(float moveAngle) {
         this.moveAngle = moveAngle;
     }
 
@@ -55,36 +55,36 @@ public class Intent {
         }
         //Angle when player press top right
         if (top == true && bottom == false && left == false && right == true) {
-            this.moveAngle = Math.toRadians(45);
+            this.moveAngle = (float)Math.toRadians(45);
         }
         //Angle when player press top
         if (top == true && bottom == false && left == false && right == false) {
-            this.moveAngle = Math.toRadians(90);
+            this.moveAngle = (float)Math.toRadians(90);
         }
         //Angle when player press top and left
         if (top == true && bottom == false && left == true && right == false) {
-            this.moveAngle = Math.toRadians(135);
+            this.moveAngle = (float)Math.toRadians(135);
         }
         //Angle when player press left
         if (left == true && bottom == false && top == false && right == false) {
-            this.moveAngle = Math.toRadians(180);
+            this.moveAngle = (float)Math.toRadians(180);
         }
         //Angle when player press left and bottom
         if (bottom == true && top == false && left == true && right == false) {
-            this.moveAngle = Math.toRadians(220);
+            this.moveAngle = (float)Math.toRadians(220);
         }
 
         //Angle when player press bottom
         if (bottom == true && top == false && left == false && right == false) {
-            this.moveAngle = Math.toRadians(270);
+            this.moveAngle = (float)Math.toRadians(270);
         }
         //Angle when player press bottom and right
         if (bottom == true && top == false && left == false && right == true) {
-            this.moveAngle = Math.toRadians(315);
+            this.moveAngle = (float)Math.toRadians(315);
         }
         //Angle when player press right
         if (right == true && bottom == false && top == false && left == false) {
-            this.moveAngle = Math.toRadians(360);
+            this.moveAngle = (float)Math.toRadians(360);
         }
